learning_web_2_bug/requirements.txt
awesome-slugify==1.6.5
Django==2.2.4
django-appconf==1.0.3
django-ckeditor==5.7.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.10.0
jieba==0.39
Markdown==3.1.1
pilkit==2.0
Pillow==6.1.0
pytz==2019.2
redis==3.3.8
regex==2019.8.19
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Unidecode==0.4.21
Whoosh==2.7.4
